Move status indicator to Camera Settings page
