+++ client/src/views/v2/CameraSettingsView.js
@@ -1,7 +1,9 @@
-import React from 'react';
 import ContentPage from 'components/ContentPage';
-import CameraSettings from 'components/v2/CameraSettings';
 import CameraRecording from 'components/v2/CameraRecording';
+import CameraSettings from 'components/v2/CameraSettings';
+import CameraStatus from 'components/v2/CameraStatus';
+import React from 'react';
+import { Card, Col, Row } from 'react-bootstrap';
 
 /**
  * Camera Settings page component
@@ -11,6 +13,19 @@ import CameraRecording from 'components/v2/CameraRecording';
 export default function CameraSettingsView() {
   return (
     <ContentPage title="Camera Settings">
+      <div className="mb-4">
+        <Card>
+          <Card.Body>
+            <Card.Title>Camera Status</Card.Title>
+            <Card.Subtitle>
+              <Row>
+                <Col><CameraStatus device="primary" /></Col>
+                <Col><CameraStatus device="secondary" /></Col>
+              </Row>
+            </Card.Subtitle>
+          </Card.Body>
+        </Card>
+      </div>
       <div className="mb-4">
         <CameraRecording devices={['primary', 'secondary']} />
       </div>
